mesa/main/framebuffer.h: Fix one -Wsign-compare warning in ?: construct.
Explicitely convert on value to the target type.

+++ src/mesa/main/framebuffer.h
@@ -93,7 +93,7 @@ static inline GLuint
 _mesa_geometric_samples(const struct gl_framebuffer *buffer)
 {
    return buffer->_HasAttachments ?
-      buffer->Visual.samples :
+      (GLuint)buffer->Visual.samples :
       buffer->DefaultGeometry._NumSamples;
 }
